+++ packages/stubee2-rolling-ui/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @stubee2/stubee2-rolling-ui
 
+## 3.1.5
+
+### Patch Changes
+
+- release
+
+## 3.1.4
+
+### Patch Changes
+
+- release
+
+## 3.1.3
+
+### Patch Changes
+
+- fix dropdownselect props
+
+## 3.1.2
+
+### Patch Changes
+
+- fix drowdownselect props
+
 ## 3.1.1
 
 ### Patch Changes